fix maxZoom configuration of Locate plugin
