+++ web/client/components/map/leaflet/index.js
@@ -1,6 +1,5 @@
 module.exports = {
     LLayer: require('./Layer.jsx'),
-    Locate: require('./Locate.jsx').default,
     LMap: require('./Map.jsx'),
     MeasurementSupport: require('./MeasurementSupport'),
     Overview: require('./Overview'),
