+++ web/client/components/map/openlayers/index.js
@@ -1,6 +1,5 @@
 import FeatureComp from './Feature';
 import LLayerComp from './Layer.jsx';
-import LocateComp from './Locate.jsx';
 import LMapComp from './Map.jsx';
 import MeasurementSupportComp from './MeasurementSupport';
 import OverviewComp from './Overview';
@@ -8,7 +7,6 @@ import ScaleBarComp from './ScaleBar';
 
 export const Feature = FeatureComp;
 export const LLayer = LLayerComp;
-export const Locate = LocateComp;
 export const LMap = LMapComp;
 export const MeasurementSupport = MeasurementSupportComp;
 export const Overview = OverviewComp;
@@ -16,7 +14,6 @@ export const ScaleBar = ScaleBarComp;
 
 export default {
     LLayer,
-    Locate,
     LMap,
     MeasurementSupport,
     Overview,
