+++ web/client/components/map/plugins/leaflet.js
@@ -12,7 +12,6 @@ export default () => {
         Map: require('../leaflet/Map'),
         Layer: require('../leaflet/Layer'),
         Feature: require('../leaflet/Feature'),
-        Locate: require('../leaflet/Locate').default,
         MeasurementSupport: require('../leaflet/MeasurementSupport'),
         Overview: require('../leaflet/Overview'),
         ScaleBar: require('../leaflet/ScaleBar'),
