+++ web/client/map/hooks/useMapTool.js
@@ -8,6 +8,11 @@
 
 import {useEffect, useRef, useState} from "react";
 
+/**
+ * Return a map tool given the map type and tool name
+ * @param  {string} mapType map type one of openlayers, leaflet or cesium
+ * @param  {string} tool map tool name
+ */
 const useMapTool = (mapType, tool) => {
     const [loaded, setLoaded] = useState(false);
     const [error, setError] = useState(null);
