+++ web/client/map/openlayers/locate.js
@@ -8,17 +8,42 @@
 
 import OlLocate from '../../utils/openlayers/OlLocate';
 
+const defaultOpt = {
+    follow: true, // follow with zoom and pan the user's location
+    remainActive: true,
+    metric: true,
+    stopFollowingOnDrag: true,
+    keepCurrentZoomLevel: false,
+    locateOptions: {
+        maximumAge: 2000,
+        enableHighAccuracy: false,
+        timeout: 10000,
+        maxZoom: 18
+    }
+};
+
+function mergeOptions(options) {
+    return {
+        ...defaultOpt,
+        ...options,
+        locateOptions: {
+            ...defaultOpt.locateOptions,
+            ...options.locateOptions
+        }
+    };
+}
 class Locate {
     start({map, options, messages, status, onStateChange, onLocationError}) {
-        this.locate = new OlLocate(map, options);
+        this.locate = new OlLocate(map, mergeOptions(options));
         this.locate.setStrings(messages);
         this.locate.options.onLocationError = onLocationError;
         this.locate.on("propertychange", (e) => {onStateChange(e.target.get(e.key)); });
         this.configureLocate(status);
     }
-    update({status, messages}) {
+    update({ status, messages, options }) {
         this.configureLocate(status);
         this.locate.setStrings(messages);
+        this.locate.setTrackingOptions(mergeOptions(options).locateOptions);
     }
     clear() {}
